YEL-139 [test] 컨트롤러 테스트를 위한 유틸 객체 생성
